backend/link-analyser-service/requirements.txt
fastapi==0.101.0
uvicorn[standard]==0.24.0
firebase-admin==6.0.0
python-dotenv==1.0.0
requests==2.31.0
python-whois==0.7.3
beautifulsoup4==4.12.2